+++ src/directives/FxOn.directive.ts
@@ -60,7 +60,7 @@ import {
     this.removeType()
     if(this.timeout){
       clearTimeout(this.timeout)
-      this.timeout = null
+      delete this.timeout
     }
   }