+++ src/directives/ShakeOn.directive.ts
@@ -67,7 +67,7 @@ import {
     this.removeType()
     if(this.timeout){
       clearTimeout(this.timeout)
-      this.timeout = null
+      delete this.timeout
     }
   }