+++ src/directives/StatusOnlineModel.directive.ts
@@ -12,10 +12,10 @@ export class StatusOnlineModel{
   @Output() statusOnlineModelChange = new EventEmitter()
 
   constructor(){
-    this.onChange = function(){
+    this.onChange = ()=>{
       this.statusOnlineModel = navigator.onLine
       this.statusOnlineModelChange.emit(this.statusOnlineModel)
-    }.bind(this)
+    }
 
     window.addEventListener("online", this.onChange)
     window.addEventListener("offline", this.onChange)