+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # ack-angular - Change Log
 All notable changes to this project will be documented here.
 
+## [1.8.2] - 2018-09-20
+- ack-sections mode default is calculate
+- strict mode build
+
+## [1.8.1] - 2018-09-20
+- ack-sections header footer height monitoring
+
 ## [1.8.0] - 2018-09-19
 - Made animate.css available
 - Added fxOn